refactor: reorganize repository into modular src/agent layout
