+++ .eslintrc.json
@@ -30,7 +30,7 @@
     "overrides": [
         {
             "files": [
-                "api/*.js"
+                "src/api/*.js"
             ],
             "parserOptions": {
                 "sourceType": "module"
@@ -41,7 +41,7 @@
         },
         {
             "files": [
-                "validate-ideas-urls.js"
+                "agent/scripts/validate-ideas-urls.js"
             ],
             "env": {
                 "node": true,
