+++ .github/workflows/tenet-pr-review.yml
@@ -29,10 +29,10 @@ jobs:
         with:
           python-version: "3.11"
           cache: "pip"
-          cache-dependency-path: .github/tenet_agent/requirements.txt
+          cache-dependency-path: agent/tenet_agent/requirements.txt
 
       - name: Install TENET agent dependencies
-        run: pip install -r .github/tenet_agent/requirements.txt
+        run: pip install -r agent/tenet_agent/requirements.txt
 
       - name: Check TENET_AI_KEY is configured
         run: |
@@ -56,5 +56,5 @@ jobs:
           BASE_SHA: ${{ github.event.pull_request.base.sha }}
           HEAD_SHA: ${{ github.event.pull_request.head.sha }}
         run: |
-          cd .github/tenet_agent
+          cd agent/tenet_agent
           python tenet_review.py
