+++ index.html
@@ -11,11 +11,11 @@
   <meta property="og:title" content="Live GSoC 2026 Org Finder & Good First Issue Tracker" />
   <meta property="og:description" content="Discover Google Summer of Code 2026 organizations and track live good first issues across GitHub." />
   <meta property="og:url" content="https://findmygsoc.vercel.app/" />
-  <meta property="og:image" content="https://findmygsoc.vercel.app/og-image.jpeg" />
+  <meta property="og:image" content="https://findmygsoc.vercel.app/src/assets/og-image.jpeg" />
   <meta name="twitter:card" content="summary_large_image" />
   <meta name="twitter:title" content="Live GSoC 2026 Org Finder & Good First Issue Tracker" />
   <meta name="twitter:description" content="Find your best-fit GSoC 2026 org and track beginner-friendly GitHub issues." />
-  <meta name="twitter:image" content="https://findmygsoc.vercel.app/og-image.jpeg" />
+  <meta name="twitter:image" content="https://findmygsoc.vercel.app/src/assets/og-image.jpeg" />
   <script type="application/ld+json">
     {
       "@context": "https://schema.org",
@@ -841,7 +841,7 @@
 </div>
 
 <!-- ═══════════════════ JS ═══════════════════ -->
-<script src="scripts/orgs.js"></script>
+<script src="agent/scripts/orgs.js"></script>
 <script>
   // ══════════════════════════════════════════════
   // THEME
@@ -1432,6 +1432,6 @@ if(org.ideas){
 </script>
 
 <!-- Vercel Speed Insights -->
-<script type="module" src="scripts/speed-insights.js"></script>
+<script type="module" src="agent/scripts/speed-insights.js"></script>
 </body>
 </html>
