+++ package.json
@@ -4,8 +4,8 @@
     "description": "GSoC 2026 Organisation Finder",
     "private": true,
     "scripts": {
-        "lint:js": "eslint app.js org.js validate-ideas-urls.js api/github.js",
-        "lint:css": "stylelint styles.css",
+        "lint:js": "eslint src/js/app.js src/js/org.js agent/scripts/validate-ideas-urls.js src/api/github.js",
+        "lint:css": "stylelint src/styles.css",
         "lint:html": "htmlhint --config .htmlhintrc index.html",
         "lint": "npm run lint:js && npm run lint:css && npm run lint:html"
     },
